Adding Wikipedia Search Engine Demo File
In this Demo Example, we create Search Engine based on Wikipedia Data. 
The User will provide Keyword ( or ) Context Phrase to search for the related information.
This Example will take the given Keyword / Phrase and filter out the Wikipedia Documents and returns the Top Matching Information. 
The solution is accelerated with Xilinx Alveo U50